auth.service

+++ frontend/src/app/services/auth.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'  // ça rend le service accessible partout
@@ -9,8 +10,14 @@ export class AuthService {
 
   // URL de ton backend
   private API = 'http://localhost:8081/api/auth';
+  private isBrowser: boolean;
 
-  constructor(private http: HttpClient) { }
+  constructor(
+    private http: HttpClient,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {
+    this.isBrowser = isPlatformBrowser(this.platformId);
+  }
 
   // --- LOGIN ---
   login(data: { email: string, password: string }): Observable<any> {
@@ -26,22 +33,32 @@ export class AuthService {
 
   // --- STOCKER TOKEN + ROLE ---
   saveSession(token: string, role: string) {
-    localStorage.setItem('token', token);
-    localStorage.setItem('role', role);
+    if (this.isBrowser) {
+      localStorage.setItem('token', token);
+      localStorage.setItem('role', role);
+    }
   }
 
   // --- RECUPERER ROLE ---
   getRole(): string | null {
-    return localStorage.getItem('role');
+    if (this.isBrowser) {
+      return localStorage.getItem('role');
+    }
+    return null;
   }
 
   // --- CHECK LOGIN ---
   isLoggedIn(): boolean {
-    return !!localStorage.getItem('token');
+    if (this.isBrowser) {
+      return !!localStorage.getItem('token');
+    }
+    return false;
   }
 
   // --- LOGOUT ---
   logout(): void {
-    localStorage.clear();
+    if (this.isBrowser) {
+      localStorage.clear();
+    }
   }
 }
